Split docker lifecycle modules and add harness command registry
