+++ rust/loopflow/src/lfd/executor/docker/types.rs
@@ -1,0 +1,2 @@
+// Lifecycle modules share Docker executor types through the parent module scope.
+// This file is reserved for future type extraction as the docker module evolves.
