+++ scratch/infra-core-boundary-cleanup.md
@@ -2,15 +2,15 @@
 
 ## Problem
 
-`lfd/store/mod.rs` (1,849 lines), `lfd/executor/docker.rs` (2,839 lines), and `engine/agent.rs` (provider command switching) are concentrated seams where small changes create large regression blast radius.
+`lfd/store/mod.rs` (1,849 lines), `lfd/executor/docker.rs` (2,839 lines), and `engine/agent.rs` (harness command switching) are concentrated seams where small changes create large regression blast radius.
 
-This pass serves maintainers and contributors first: faster reviews, safer refactors, and simpler provider additions. Users benefit indirectly through fewer orchestration regressions and more predictable wave execution.
+This pass serves maintainers and contributors first: faster reviews, safer refactors, and simpler harness additions. Users benefit indirectly through fewer orchestration regressions and more predictable wave execution.
 
 Why now: wave pass 1 is explicitly the deconcentration pass before contract-hardening and orchestration expansion. If we skip this, later features compound complexity instead of leverage.
 
 Wave-goal alignment:
 - Advances **"Eliminate boilerplate and duplicated patterns"** by removing store forwarder/match repetition.
-- Advances **"Make extension points trait-based, not switch-based"** by replacing provider command branching with provider-owned builders.
+- Advances **"Make extension points trait-based, not switch-based"** by replacing harness command branching with harness-owned builders.
 - Advances **"Maintain architectural compactness as features grow"** by splitting `docker` responsibilities into lifecycle modules.
 
 ## Approach
@@ -48,21 +48,23 @@ Rules:
 
 Concrete target: no single docker module exceeds ~900 lines.
 
-### 3) Provider command registry: provider-owned builders
+### 3) Harness command registry: harness-owned builders
 
-- Create `engine/provider_commands/` with a `ProviderCommandBuilder` trait:
+- Create `engine/harness_commands/` with a `HarnessCommandBuilder` trait:
   - `build_command(...) -> Vec<String>`
   - `apply_env(...)`
-- Move provider-specific command logic into modules:
+- Move harness-specific command logic into modules:
   - `claude.rs`, `codex.rs`, `gemini.rs`, `opencode.rs`
-- Add a registry lookup keyed by parsed backend string; `build_model_command()` becomes a thin delegator.
-- Keep unknown-model behavior parity through an explicit fallback builder (current fallback-to-Claude behavior).
+- Add a registry lookup keyed by parsed harness/backend string; `build_model_command()` becomes a thin delegator.
+- Keep unknown-model behavior parity through an explicit fallback builder: unknown models currently default to Claude and pass the full model string as the variant (current `FallbackClaude` behavior).
+- Keep the fallback harness choice isolated behind the fallback builder so we can change the default in a later pass without reintroducing central switch logic.
+- Standardize terminology on **harness** across engine and sessions where possible. Keep existing persisted field names (for example `provider_session_id`) unchanged in this pass.
 
-Concrete target: adding a provider requires adding a provider module + registration entry, with no central `match` branch growth.
+Concrete target: adding a harness requires adding a harness module + registration entry, with no central `match` branch growth.
 
 ### 4) Mandatory sweep pass after each decomposition
 
-After store, provider, and docker changes, run stale-reference sweeps (code + docs + tests) to catch rename fallout early.
+After store, harness, and docker changes, run stale-reference sweeps (code + docs + tests) to catch rename fallout early.
 
 ## Alternatives considered
 
@@ -70,7 +72,7 @@ After store, provider, and docker changes, run stale-reference sweeps (code + do
 |----------|----------|---------|
 | Macro-generate forwarding in `store/mod.rs` and keep current shape | Fewer handwritten lines, but responsibility concentration remains and session dispatch inconsistency survives | Hides pain instead of fixing boundaries |
 | Full rewrite of store/executor APIs in one big migration | Clean slate architecture | High semantic risk; violates pass contract of structural-not-semantic change |
-| Keep central provider switch and only extract helper functions | Small diff, easy review | Fails wave goal to make extension points trait/registry-based; central branching pressure remains |
+| Keep central harness switch and only extract helper functions | Small diff, easy review | Fails wave goal to make extension points trait/registry-based; central branching pressure remains |
 
 ## Key decisions
 
@@ -78,13 +80,13 @@ After store, provider, and docker changes, run stale-reference sweeps (code + do
   - Why: boilerplate is the problem, not formatting.
 - **Decision: split docker by lifecycle ownership, not by technical primitive (git/docker/fs).**
   - Why: lifecycle seams map to failure domains (image/workspace/recovery/IO).
-- **Decision: provider modules own command construction and env wiring.**
-  - Why: provider-specific behavior should evolve in provider-specific files.
+- **Decision: harness modules own command construction and env wiring.**
+  - Why: harness-specific behavior should evolve in harness-specific files.
 - **Decision: enforce sweep passes as part of implementation, not optional cleanup.**
   - Why: prior direction-taxonomy work proved stale-reference blast radius is real.
 
 Wild success we are designing for:
-- new provider onboarding is a self-contained file addition
+- new harness onboarding is a self-contained file addition
 - docker recovery bugs are fixed in `recovery.rs` without touching workspace/image code
 - store changes stop causing multi-domain merge conflicts in one hotspot file
 
@@ -98,7 +100,7 @@ Wild failure we are designing against:
 - In scope:
   - Store backend port + session trait unification + call-site migration
   - Docker lifecycle module split with unchanged executor contract
-  - Provider command registry and provider-owned command/env builders
+  - Harness command registry and harness-owned command/env builders
   - Stale-reference sweep passes after each structural move
 - Out of scope:
   - Prompt budgeting or prompt rendering behavior changes
@@ -118,4 +120,4 @@ Wild failure we are designing against:
 And these observable outcomes hold:
 - `store/mod.rs` no longer carries forwarding-heavy normal call paths (wave goal: **"Eliminate boilerplate and duplicated patterns"**).
 - Docker executor logic is split into lifecycle-owned modules with `AgentExecutor` behavior parity (wave goal: **"Maintain architectural compactness as features grow"**).
-- Provider command construction is registry/trait based, not switch growth in central code (wave goal: **"Make extension points trait-based, not switch-based"**).
+- Harness command construction is registry/trait based, not switch growth in central code (wave goal: **"Make extension points trait-based, not switch-based"**).
